% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{item_weighting}
\alias{item_weighting}
\title{Weighting items via Kilic & Dogan's method}
\usage{
item_weighting(x)
}
\arguments{
\item{x}{The dataframe. Columns show items and rows show individuals.}
}
\value{
Weighted scores as data frame
}
\description{
Weighting items via Kilic & Dogan's method
}
\examples{
set.seed(1234)
data <- psych::sim.VSS(1000,16,1,0.5,dichot = TRUE)
weighted_scores <- item_weighting(data)
weighted_scores
}
